+++ src/Geohash.SpatialIndex.SpatialRelations/SpatialRelations.cs
@@ -16,12 +16,7 @@ namespace Geohash.SpatialIndex.SpatialRelations
 			_geohashSpatialIndex = geohashSpatialIndex;
 		}
 
-		/// <summary>
-		/// A simple k-nearest-neighbour search
-		/// </summary>
-		/// <param name="geom"></param>
-		/// <param name="k"></param>
-		/// <returns></returns>
+		///<inheritdoc/>
 		public IEnumerable<IndexEntry<T>> KNN(int k, Geometry geom,
 			Func<IndexEntry<T>, double> distanceMetric,
 			T exclude = default)
@@ -37,11 +32,7 @@ namespace Geohash.SpatialIndex.SpatialRelations
 			return result.OrderBy(g => distanceMetric(g)).Take(k).ToList();
 		}
 
-		/// <summary>
-		/// Returns all geometries in the index that contains the given geometry
-		/// </summary>
-		/// <param name="geom"></param>
-		/// <returns></returns>
+		///<inheritdoc/>
 		public IEnumerable<IndexEntry<T>> STContains(Geometry geom, T exclude = default)
 		{
 			var (trieMapSearchResult, _) = _geohashSpatialIndex.Query(geom, minimumHits:0, exclude: exclude);
@@ -49,11 +40,7 @@ namespace Geohash.SpatialIndex.SpatialRelations
 			return trieMapSearchResult.Where(ie => preparedGeom.Contains(ie.Geom)).ToList();
 		}
 
-		/// <summary>
-		/// Returns all geometries in the index that properly contains the given geometry
-		/// </summary>
-		/// <param name="geom"></param>
-		/// <returns></returns>
+		///<inheritdoc/>
 		public IEnumerable<IndexEntry<T>> STContainsProperly(Geometry geom, T exclude = default)
 		{
 			var (trieMapSearchResult, _) = _geohashSpatialIndex.Query(geom, minimumHits: 0, exclude: exclude);
@@ -61,11 +48,7 @@ namespace Geohash.SpatialIndex.SpatialRelations
 			return trieMapSearchResult.Where(ie => preparedGeom.ContainsProperly(ie.Geom)).ToList();
 		}
 
-		/// <summary>
-		/// Returns all geometries in the index that crosses with the given geometry
-		/// </summary>
-		/// <param name="geom"></param>
-		/// <returns></returns>
+		///<inheritdoc/>
 		public IEnumerable<IndexEntry<T>> STCrosses(Geometry geom, T exclude = default)
 		{
 			var (trieMapSearchResult, _) = _geohashSpatialIndex.Query(geom, exclude: exclude);
@@ -73,11 +56,7 @@ namespace Geohash.SpatialIndex.SpatialRelations
 			return trieMapSearchResult.Where(ie => preparedGeom.Crosses(ie.Geom)).ToList();
 		}
 
-		/// <summary>
-		/// Returns all geometries in the index have an exactly equal geometry
-		/// </summary>
-		/// <param name="geom"></param>
-		/// <returns></returns>
+		///<inheritdoc/>
 		public IEnumerable<IndexEntry<T>> STEquals(Geometry geom, T exclude = default)
 		{
 			var (trieMapSearchResult, _) = _geohashSpatialIndex.Query(geom, minimumHits: 0, exclude: exclude);
@@ -85,11 +64,7 @@ namespace Geohash.SpatialIndex.SpatialRelations
 			return trieMapSearchResult.Where(ie => preparedGeom.Equals(ie.Geom)).ToList();
 		}
 
-		/// <summary>
-		/// Returns all geometries in the index that intersects with the given geometry
-		/// </summary>
-		/// <param name="geom"></param>
-		/// <returns></returns>
+		///<inheritdoc/>
 		public IEnumerable<IndexEntry<T>> STIntersects(Geometry geom, T exclude = default)
 		{
 			var (trieMapSearchResult, _) = _geohashSpatialIndex.Query(geom, exclude: exclude);
@@ -104,11 +79,7 @@ namespace Geohash.SpatialIndex.SpatialRelations
 			}
 		}
 
-		/// <summary>
-		/// Returns the closest geometry from the index
-		/// </summary>
-		/// <param name="geom"></param>
-		/// <returns></returns>
+		///<inheritdoc/>
 		public IndexEntry<T> STNearestNeighbour(Geometry geom,
 			Func<IndexEntry<T>, double> distanceMetric,
 			T exclude = default)
@@ -117,11 +88,7 @@ namespace Geohash.SpatialIndex.SpatialRelations
 			return trieMapSearchResult.OrderBy(g => distanceMetric(g)).FirstOrDefault();
 		}
 
-		/// <summary>
-		/// Returns all geometries in the index that overlaps with the given geometry
-		/// </summary>
-		/// <param name="geom"></param>
-		/// <returns></returns>
+		///<inheritdoc/>
 		public IEnumerable<IndexEntry<T>> STOverlaps(Geometry geom, T exclude = default)
 		{
 			var (trieMapSearchResult, _) = _geohashSpatialIndex.Query(geom, exclude: exclude);
@@ -129,11 +96,7 @@ namespace Geohash.SpatialIndex.SpatialRelations
 			return trieMapSearchResult.Where(ie => preparedGeom.Overlaps(ie.Geom)).ToList();
 		}
 
-		/// <summary>
-		/// Returns all geometries in the index that touches the given geometry
-		/// </summary>
-		/// <param name="geom"></param>
-		/// <returns></returns>
+		///<inheritdoc/>
 		public IEnumerable<IndexEntry<T>> STTouches(Geometry geom, T exclude = default)
 		{
 			var (trieMapSearchResult, _) = _geohashSpatialIndex.Query(geom, exclude: exclude);
@@ -141,12 +104,7 @@ namespace Geohash.SpatialIndex.SpatialRelations
 			return trieMapSearchResult.Where(ie => preparedGeom.Touches(ie.Geom)).ToList();
 		}
 
-		/// <summary>
-		/// Returns all geometries in the index that is within the given geometry
-		/// </summary>
-		/// <param name="geom"></param>
-		/// <param name="exclude"></param>
-		/// <returns></returns>
+		///<inheritdoc/>
 		public IEnumerable<IndexEntry<T>> STWithin(Geometry geom, T exclude = default)
 		{
 			var (trieMapSearchResult, hash) = _geohashSpatialIndex.Query(geom, minimumHits:0, exclude: exclude);