+++ src/Geohash.SpatialIndex.Tests/GeohashIndexUpdateTests.cs
@@ -1,0 +1,63 @@
+﻿using Geohash.SpatialIndex.Core;
+using Geohash.SpatialIndex.SpatialRelations;
+using NetTopologySuite.Geometries;
+using NUnit.Framework;
+using System.Linq;
+
+namespace Geohash.SpatialIndex.Tests
+{
+	public class GeohashIndexUpdateTests
+	{
+		protected GeohashSpatialIndex<int> emptyIndex;
+
+		[SetUp]
+		public void Setup()
+		{
+			emptyIndex = new GeohashSpatialIndex<int>(new DefaultGeohasher(), new DefaultTrieMap<int>(), 9);
+		}
+
+		[Test]
+		public void TestRemoveSingleEntryByValue()
+		{
+			emptyIndex.Insert(new Point(1, 2), 1);
+			emptyIndex.Insert(new Point(2, 1), 2);
+			emptyIndex.Remove(1);
+			var result = emptyIndex.Query(string.Empty).Entries;
+			Assert.IsTrue(result.Count() == 1);
+			Assert.AreEqual(2, result.Single().Value);
+		}
+
+		[Test]
+		public void TestRemoveSingleEntryOnSameLocationByValue()
+		{
+			emptyIndex.Insert(new Point(1, 2), 1);
+			emptyIndex.Insert(new Point(1, 2), 2);
+
+			var firstResult = emptyIndex.Query(string.Empty).Entries;
+			Assert.IsTrue(firstResult.Count() == 2);
+			Assert.AreEqual(1, firstResult.First().Value);
+			Assert.AreEqual(2, firstResult.Skip(1).First().Value);
+
+			emptyIndex.Remove(1);
+
+			var secondResult = emptyIndex.Query(string.Empty).Entries;
+			Assert.IsTrue(secondResult.Count() == 1);
+			Assert.AreEqual(2, secondResult.First().Value);
+		}
+
+
+		[Test]
+		public void TestInsertOrUpdate()
+		{
+			emptyIndex.InsertOrUpdate(new Point(1, 2), 1);
+			emptyIndex.InsertOrUpdate(new Point(2, 1), 1);
+			emptyIndex.InsertOrUpdate(new Point(1, 2), 2);
+			emptyIndex.InsertOrUpdate(new Point(2, 1), 2);
+
+			var result = emptyIndex.Query(string.Empty).Entries;
+			Assert.IsTrue(result.Count() == 2);
+			Assert.AreEqual(1, result.First().Value);
+			Assert.AreEqual(2, result.Skip(1).First().Value);
+		}
+	}
+}