+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## Version 0.2.0
+* Added Remove(T value) method to remove entries by value
+* Added InsertOrUpdate(geohash, value) method. Note that the current implementation does a simple Remove(value) and then Insert(geohash, value), which is slow. Avoid using this when building the index from bulk data.
+* Updated NuGet package dependencies
+
 ## Version 0.1.1-pre
 * Added license expression to NuGet packages
 * Updated Authors field and added link to release notes